Lint

+++ client/allocrunner/taskrunner/task_runner.go
@@ -706,8 +706,8 @@ func (tr *TaskRunner) buildTaskConfig() *drivers.TaskConfig {
 		Resources: &drivers.Resources{
 			NomadResources: taskResources,
 			LinuxResources: &drivers.LinuxResources{
-				MemoryLimitBytes: int64(taskResources.Memory.MemoryMB) * 1024 * 1024,
-				CPUShares:        int64(taskResources.Cpu.CpuShares),
+				MemoryLimitBytes: taskResources.Memory.MemoryMB * 1024 * 1024,
+				CPUShares:        taskResources.Cpu.CpuShares,
 				PercentTicks:     float64(taskResources.Cpu.CpuShares) / float64(tr.clientConfig.Node.NodeResources.Cpu.CpuShares),
 			},
 		},
